organizes files into folders
