+++ src/commands/tankdrive.py
@@ -1,8 +1,11 @@
-
+import sys, os, inspect
+currentdir = os.path.dirname(os.path.abspath(inspect.getfile(inspect.currentframe())))
+parentdir = os.path.dirname(currentdir)
+sys.path.insert(0,parentdir) 
 from wpilib.command import Command
 import math
 import oi
-import drive
+from subsystems import drive
 
 class TankDrive(Command):
     def __init__(self):
